refactor(parser): split paragraphs/schemas/proofs into parser_pkg/ submodules
Phase 1B Batch 6 (Moves 12-14 from docs/PHASE1-PLAN.md).  Begin
the parser split: extract paragraph, schema, and proof parser rules
from parser.py into three new mixin modules under
src/txt2tex/parser_pkg/.

Setup (analog of Move 0):
* parser_pkg/_base.py — defines ParserBase, a type-only shape
  declaration that mirrors the Parser interface.  Mixins inherit
  from it so their method bodies type-check against the full
  parser surface area without importing parser.py itself
  (circular).  Method-body discipline is identical to codegen:
  byte-identical Fowler Move Method.
* parser_pkg/_base.py also holds ParserError, relocated from
  parser.py.  parser.py re-exports the name (``from ... import
  ParserError as ParserError``) for backward compatibility.

Move 12 — parser_pkg/paragraphs.py
  Rules for: GivenType, FreeType (and FreeBranch helper),
  SyntaxBlock, Abbreviation, AxDef, GenDef, Zed.  Plus the
  ``_parse_compound_identifier_name`` helper consumed only by
  ``_parse_abbreviation``.

Move 13 — parser_pkg/schemas.py
  Rules for: Schema, HorizDef (with helpers
  ``_parse_horiz_def_rhs`` and ``_parse_horiz_def_generic_params``),
  ``_parse_schema_text`` (the inline ``[decl | pred]`` form), the
  three schema-calculus operators
  (``_parse_schema_pipe / _compose / _project_hide``), and the
  bracket-dispatch entries ``_parse_schema_rename_or_generic`` /
  ``_parse_schema_rename``.

Move 14 — parser_pkg/proofs.py
  Rules for: ProofTree, ProofNode, CaseAnalysis, ArgueChain
  (ARGUE / EQUIV / EQUAL), InfruleBlock, plus the
  ``_smart_join_justification`` helper shared by argue/infrule/proof.

pyproject.toml — parser_pkg/*.py joins the ARG002 allowlist
(unused parent_indent in proof-tree parsing, same precedent as
parser.py).

Gate 6 verified:
* make check — lint + markdown + format + mypy + pyright + 4700
  unit tests, all clean.
* make test-e2e — 159/159 examples fixtures byte-identical.
* make refactor-diff — 190/190 byte-identical vs main.

parser.py shrinks from 6367 → 4666 lines (cumulative 26.7%).
