+++ src/txt2tex/parser.py
@@ -7,18 +7,13 @@ import re
 from typing import ClassVar, Literal
 
 from txt2tex.ast_nodes import (
-    Abbreviation,
     Aggregator,
     AggregatorClause,
-    ArgueChain,
-    ArgueStep,
-    AxDef,
     BagLiteral,
     BibliographyMetadata,
     BinaryOp,
     Binding,
     BMachine,
-    CaseAnalysis,
     Conditional,
     Contents,
     Declaration,
@@ -26,20 +21,14 @@ from txt2tex.ast_nodes import (
     Document,
     DocumentItem,
     Expr,
-    FreeBranch,
-    FreeType,
     FunctionApp,
     FunctionType,
-    GenDef,
     GenericInstantiation,
-    GivenType,
     Group,
     GroupAggregate,
     GuardedBranch,
     GuardedCases,
-    HorizDef,
     Identifier,
-    InfruleBlock,
     Lambda,
     LatexBlock,
     LineBreak,
@@ -50,8 +39,6 @@ from txt2tex.ast_nodes import (
     Part,
     PartsFormat,
     Project,
-    ProofNode,
-    ProofTree,
     PureParagraph,
     Quantifier,
     Range,
@@ -59,15 +46,8 @@ from txt2tex.ast_nodes import (
     RelationalImage,
     RelationRename,
     Restrict,
-    Schema,
     SchemaBinding,
-    SchemaCompose,
-    SchemaHide,
     SchemaInclusion,
-    SchemaPipe,
-    SchemaProject,
-    SchemaRename,
-    SchemaText,
     Section,
     SequenceLiteral,
     SetComprehension,
@@ -76,8 +56,6 @@ from txt2tex.ast_nodes import (
     StringLit,
     Subscript,
     Superscript,
-    SyntaxBlock,
-    SyntaxDefinition,
     Theta,
     TitleMetadata,
     TruthTable,
@@ -85,27 +63,27 @@ from txt2tex.ast_nodes import (
     TupleProjection,
     UnaryOp,
     Ungroup,
-    Zed,
 )
 from txt2tex.constants import PROSE_WORDS
+from txt2tex.parser_pkg._base import ParserBase, ParserError as ParserError
+from txt2tex.parser_pkg.paragraphs import (
+    _ParagraphsParser,  # pyright: ignore[reportPrivateUsage]
+)
+from txt2tex.parser_pkg.proofs import (
+    _ProofsParser,  # pyright: ignore[reportPrivateUsage]
+)
+from txt2tex.parser_pkg.schemas import (
+    _SchemasParser,  # pyright: ignore[reportPrivateUsage]
+)
 from txt2tex.tokens import Token, TokenType
 
 
-class ParserError(Exception):
-    """Raised when parser encounters invalid syntax."""
-
-    # Instance variable type annotations
-    message: str
-    token: Token
-
-    def __init__(self, message: str, token: Token) -> None:
-        """Initialize parser error with token position."""
-        super().__init__(f"Line {token.line}, column {token.column}: {message}")
-        self.message = message
-        self.token = token
-
-
-class Parser:
+class Parser(
+    _ParagraphsParser,
+    _SchemasParser,
+    _ProofsParser,
+    ParserBase,
+):
     """Recursive descent parser for txt2tex whiteboard notation.
 
     Expression grammar (precedence from lowest to highest):
@@ -1144,23 +1122,6 @@ class Parser:
                 return False
             offset += 1
 
-    def _parse_schema_rename_or_generic(
-        self, base: Expr
-    ) -> GenericInstantiation | SchemaRename:
-        """Consume '[' then dispatch to rename or generic instantiation.
-
-        Disambiguation rule (Z RM §3.11): scan the bracket contents at
-        bracket depth 0.  If any SLASH token appears before the matching
-        ']', parse as schema rename pairs ``S[new/old, ...]``.  Otherwise,
-        parse as a generic type instantiation ``S[T, ...]``.
-
-        The '[' has NOT been consumed on entry; this method consumes it and
-        the matching ']'.
-        """
-        if self._bracket_contains_slash():
-            return self._parse_schema_rename(base)
-        return self._parse_generic_instantiation(base)
-
     def _parse_generic_instantiation(self, base: Expr) -> GenericInstantiation:
         """Parse generic instantiation S[T, ...] — '[' not yet consumed."""
         lbracket_token = self._advance()  # Consume '['
@@ -1193,106 +1154,6 @@ class Parser:
             type_params=type_params,
             line=lbracket_token.line,
             column=lbracket_token.column,
-        )
-
-    def _parse_schema_rename(self, schema: Expr) -> SchemaRename:
-        """Parse schema rename S[old/new, ...] — '[' not yet consumed.
-
-        Grammar:
-            rename ::= '[' pair (',' pair)* ']'
-            pair   ::= IDENTIFIER '/' IDENTIFIER
-
-        Raises ParserError for:
-        - Empty bracket ``S[]``
-        - Missing source name ``S[/b]``
-        - Missing slash ``S[a b]``
-        - Missing target name ``S[a/]``
-        - Trailing slash ``S[a/b/]``
-        - Trailing comma ``S[a/b,]``
-        """
-        open_tok = self._advance()  # Consume '['
-
-        pairs: list[tuple[str, str]] = []
-
-        while not self._match(TokenType.RBRACKET):
-            if self._at_end() or self._match(TokenType.NEWLINE, TokenType.EOF):
-                raise ParserError(
-                    "unclosed '[' in schema rename",
-                    open_tok,
-                )
-
-            # Expect source identifier
-            if not self._match(TokenType.IDENTIFIER):
-                cur = self._current()
-                raise ParserError(
-                    f"expected identifier before '/' in schema rename pair,"
-                    f" got {cur.type.name} ({cur.value!r})",
-                    cur,
-                )
-            src_tok = self._advance()
-            src = src_tok.value
-
-            # Expect '/'
-            if not self._match(TokenType.SLASH):
-                cur = self._current()
-                raise ParserError(
-                    f"expected '/' in schema rename pair after {src!r},"
-                    f" got {cur.type.name} ({cur.value!r})",
-                    cur,
-                )
-            self._advance()  # Consume '/'
-
-            # Expect target identifier
-            if not self._match(TokenType.IDENTIFIER):
-                cur = self._current()
-                raise ParserError(
-                    f"expected identifier after '/' in schema rename pair,"
-                    f" got {cur.type.name} ({cur.value!r})",
-                    cur,
-                )
-            tgt_tok = self._advance()
-            tgt = tgt_tok.value
-
-            pairs.append((src, tgt))
-
-            # Optional comma separator
-            if self._match(TokenType.COMMA):
-                self._advance()  # Consume ','
-                # Trailing comma check: next is ']'
-                if self._match(TokenType.RBRACKET):
-                    raise ParserError(
-                        "expected rename pair after ',' in schema rename",
-                        self._current(),
-                    )
-                continue
-
-            # If not comma and not ']', something is wrong
-            if not self._match(TokenType.RBRACKET):
-                cur = self._current()
-                raise ParserError(
-                    f"expected ',' or ']' in schema rename,"
-                    f" got {cur.type.name} ({cur.value!r})",
-                    cur,
-                )
-
-        if not pairs:
-            raise ParserError(
-                "schema rename requires at least one pair",
-                open_tok,
-            )
-
-        if not self._match(TokenType.RBRACKET):
-            raise ParserError(
-                "expected ']' to close schema rename",
-                self._current(),
-            )
-        self._advance()  # Consume ']'
-
-        return SchemaRename(
-            schema=schema,
-            pairs=pairs,
-            line=open_tok.line,
-            column=open_tok.column,
         )
 
     def _parse_declaration_or_inclusion(
@@ -1446,39 +1307,6 @@ class Parser:
             )
             for vt in var_tokens
         ]
-
-    def _smart_join_justification(self, parts: list[str]) -> str:
-        """Join justification tokens intelligently.
-
-        Strategy: Join with spaces (preserving word separation), then remove
-        spaces around parentheses, brackets, and equals signs to compact
-        mathematical notation.
-
-        Examples:
-        - ["length", "(", "<", "ple", ">", "^", "pl", ")"] → "length(<ple>^pl)"
-        - ["commutativity", "of", "or"] → "commutativity of or"
-        - ["x", "is", "not", "free", "in", "y", "|->", "z"] → "x is not free in y|->z"
-        - ["def", ".", "of", "=>"] → "def. of =>" (preserves space before =>)
-        """
-        if not parts:
-            return ""
-
-        # First join with spaces (preserves all word separation)
-        result = " ".join(parts)
-
-        # Remove spaces around specific punctuation to compact notation
-        # ONLY touch safe characters that aren't part of operators
-        result = re.sub(r"\s*\(\s*", "(", result)  # Remove space before/after (
-        result = re.sub(r"\s*\)\s*", ")", result)  # Remove space before/after )
-        # Remove space around standalone = but NOT part of operators (=>, <=>)
-        # Negative lookahead/lookbehind avoids matching = in operators
-        result = re.sub(r"\s*(?<![<>])=(?![>])\s*", "=", result)
-        result = re.sub(r"\s*\.", ".", result)  # Remove space BEFORE period only
-        return re.sub(r"\s*,", ",", result)  # Remove space BEFORE comma only
-
-        # DO NOT touch < or > as they appear in many operators (=>, ->>, etc.)
-        # Sequences like <ple> will have internal spacing, but that's acceptable
-        # to avoid breaking operator conversion in _escape_justification
 
     def _parse_section(self) -> Section:
         """Parse section: === Title ==="""
@@ -1665,207 +1493,6 @@ class Parser:
         return TruthTable(
             headers=headers,
             rows=rows,
-            line=start_token.line,
-            column=start_token.column,
-        )
-
-    def _parse_argue_chain(self, connector: Literal["iff", "eq"] = "iff") -> ArgueChain:
-        """Parse ARGUE:, EQUIV:, or EQUAL: block into an ArgueChain.
-
-        Both EQUIV: and ARGUE: keywords produce connector='iff', rendered
-        with \\Leftrightarrow between steps. EQUAL: produces connector='eq',
-        rendered with = between steps (for expression equality chains).
-        The generator uses a standard LaTeX array environment — not fuzz's
-        argue environment — for reliable column spacing and adjustbox scaling.
-        """
-        start_token = self._advance()  # Consume 'ARGUE:', 'EQUIV:', or 'EQUAL:'
-        self._skip_newlines()
-
-        steps: list[ArgueStep] = []
-
-        # Parse steps until we hit another structural element or end
-        while not self._at_end() and not self._is_structural_token():
-            # Consume leading connective if present on continuation lines.
-            # Connector specificity: an EQUAL: block must never silently consume
-            # a leading <=>, and an EQUIV/ARGUE block must never consume a leading =.
-            # Note: = on a continuation line is consumed by _parse_comparison
-            # (which crosses newlines), so the EQUALS branch rarely fires in
-            # practice. The IFF branch fires because _parse_iff does NOT cross
-            # newlines, so a bare <=> at line start reaches this check.
-            if connector == "eq" and self._match(TokenType.EQUALS):
-                self._advance()  # Consume leading '='
-            elif connector == "iff" and self._match(TokenType.IFF):
-                self._advance()  # Consume leading '<=>'
-            elif connector == "eq" and self._match(TokenType.IFF):
-                # <=> at the start of an EQUAL: continuation line means the
-                # author used the wrong block type.  A clear message is more
-                # helpful than the generic "unexpected token" from _parse_expr.
-                tok = self._current()
-                msg = (
-                    "found '<=>' continuation in an EQUAL: chain; "
-                    "use EQUIV: or ARGUE: for logical-equivalence chains"
-                )
-                raise ParserError(msg, tok)
-            elif connector == "iff" and self._match(TokenType.EQUALS):
-                # bare '=' at the start of an EQUIV:/ARGUE: continuation line
-                # — the author may have intended EQUAL:.  Note: in practice
-                # _parse_comparison absorbs 'a\n= b' as a single step, so
-                # this branch fires only for a truly bare '=' with nothing
-                # before it on the line, which is syntactically invalid.
-                tok = self._current()
-                msg = (
-                    "found '=' continuation in an EQUIV:/ARGUE: chain; "
-                    "use EQUAL: for expression-equality chains"
-                )
-                raise ParserError(msg, tok)
-
-            # Parse expression
-            expr = self._parse_expr()
-            self._skip_newlines()
-
-            # Check for optional justification [text]
-            justification: str | None = None
-            if self._match(TokenType.LBRACKET):
-                self._advance()  # Consume '['
-
-                # Collect justification text (all tokens until ']')
-                just_parts: list[str] = []
-                while not self._match(TokenType.RBRACKET) and not self._at_end():
-                    if self._match(TokenType.NEWLINE):
-                        break  # Justification on one line
-                    just_parts.append(self._current().value)
-                    self._advance()
-
-                if not self._match(TokenType.RBRACKET):
-                    raise ParserError(
-                        "Expected closing ']' for justification", self._current()
-                    )
-
-                self._advance()  # Consume ']'
-                # Join tokens smartly: add spaces only between consecutive identifiers
-                justification = self._smart_join_justification(just_parts)
-
-            # Create step
-            steps.append(
-                ArgueStep(
-                    expression=expr,
-                    justification=justification,
-                    line=expr.line,
-                    column=expr.column,
-                )
-            )
-
-            self._skip_newlines()
-
-        if not steps:
-            raise ParserError(
-                "Expected at least one step in ARGUE/EQUIV/EQUAL chain",
-                self._current(),
-            )
-
-        return ArgueChain(
-            steps=steps,
-            connector=connector,
-            line=start_token.line,
-            column=start_token.column,
-        )
-
-    def _parse_infrule_block(self) -> InfruleBlock:
-        """Parse INFRULE: block with premises, ---, and conclusion.
-
-        Format:
-          INFRULE:
-          premise1 [label1]
-          premise2 [label2]
-          ---
-          conclusion [label]
-        """
-        start_token = self._advance()  # Consume 'INFRULE:'
-        self._skip_newlines()
-
-        premises: list[tuple[Expr, str | None]] = []
-
-        # Parse premises until we hit ---
-        while not self._at_end() and not self._match(TokenType.DERIVE):
-            # Check if we hit another structural token (error)
-            if self._is_structural_token():
-                raise ParserError(
-                    "Expected '---' separator in INFRULE block", self._current()
-                )
-
-            # Parse premise expression
-            expr = self._parse_expr()
-            self._skip_newlines()
-
-            # Check for optional label [text]
-            label: str | None = None
-            if self._match(TokenType.LBRACKET):
-                self._advance()  # Consume '['
-
-                # Collect label text (all tokens until ']')
-                label_parts: list[str] = []
-                while not self._match(TokenType.RBRACKET) and not self._at_end():
-                    if self._match(TokenType.NEWLINE):
-                        break  # Label on one line
-                    label_parts.append(self._current().value)
-                    self._advance()
-
-                if not self._match(TokenType.RBRACKET):
-                    raise ParserError("Expected closing ']' for label", self._current())
-
-                self._advance()  # Consume ']'
-                label = self._smart_join_justification(label_parts)
-
-            premises.append((expr, label))
-            self._skip_newlines()
-
-        # Check for --- separator
-        if not self._match(TokenType.DERIVE):
-            raise ParserError(
-                "Expected '---' separator in INFRULE block", self._current()
-            )
-
-        self._advance()  # Consume '---'
-        self._skip_newlines()
-
-        # Parse conclusion expression
-        if self._at_end() or self._is_structural_token():
-            raise ParserError(
-                "Expected conclusion after '---' in INFRULE block", self._current()
-            )
-
-        conclusion_expr = self._parse_expr()
-        self._skip_newlines()
-
-        # Check for optional label [text]
-        conclusion_label: str | None = None
-        if self._match(TokenType.LBRACKET):
-            self._advance()  # Consume '['
-
-            # Collect label text
-            conclusion_label_parts: list[str] = []
-            while not self._match(TokenType.RBRACKET) and not self._at_end():
-                if self._match(TokenType.NEWLINE):
-                    break
-                conclusion_label_parts.append(self._current().value)
-                self._advance()
-
-            if not self._match(TokenType.RBRACKET):
-                raise ParserError(
-                    "Expected closing ']' for conclusion label", self._current()
-                )
-
-            self._advance()  # Consume ']'
-            conclusion_label = self._smart_join_justification(conclusion_label_parts)
-
-        if not premises:
-            raise ParserError(
-                "Expected at least one premise in INFRULE block", self._current()
-            )
-
-        return InfruleBlock(
-            premises=premises,
-            conclusion=(conclusion_expr, conclusion_label),
             line=start_token.line,
             column=start_token.column,
         )
@@ -4976,305 +4603,6 @@ class Parser:
             column=lbag_token.column,
         )
 
-    def _parse_given_type(self) -> GivenType:
-        """Parse given type: given A, B, C"""
-        start_token = self._advance()  # Consume 'given'
-
-        names: list[str] = []
-
-        # Parse comma-separated list of type names
-        # Stop at NEWLINE, END (for zed blocks), or EOF
-        while not self._at_end() and not self._match(TokenType.NEWLINE, TokenType.END):
-            if self._match(TokenType.COMMA):
-                self._advance()  # Skip comma
-                continue
-
-            if not self._match(TokenType.IDENTIFIER):
-                raise ParserError(
-                    "Expected type name in given declaration", self._current()
-                )
-
-            names.append(self._current().value)
-            self._advance()
-
-        if not names:
-            raise ParserError(
-                "Expected at least one type name in given declaration", self._current()
-            )
-
-        return GivenType(names=names, line=start_token.line, column=start_token.column)
-
-    def _parse_free_branch(self, branch_token: Token) -> FreeBranch:
-        """Parse a single free type branch with optional parameters.
-
-        Assumes the branch name token has already been consumed. Parses
-        optional constructor parameters in angle brackets: <...> or ⟨...⟩
-
-        Args:
-            branch_token: The token containing the branch name
-
-        Returns:
-            FreeBranch with name and optional parameters
-        """
-        branch_name = branch_token.value
-
-        # Check for constructor parameters: ⟨...⟩ or <...>
-        parameters: Expr | None = None
-        if self._match(TokenType.LANGLE):
-            langle_token = self._current()
-            self._advance()  # Consume '⟨' or '<'
-
-            # Check for empty parameters: ⟨⟩
-            if self._match(TokenType.RANGLE):
-                # Empty parameters - create empty sequence literal
-                parameters = SequenceLiteral(
-                    elements=[],
-                    line=langle_token.line,
-                    column=langle_token.column,
-                )
-                self._advance()  # Consume '⟩' or '>'
-            else:
-                # Parse parameter type expression (can be cross product)
-                # Examples: ⟨N⟩, ⟨Tree⟩, ⟨Tree x Tree⟩
-                parameters = self._parse_cross()
-
-                # Expect closing angle bracket
-                if not self._match(TokenType.RANGLE):
-                    raise ParserError(
-                        "Expected '⟩' or '>' after constructor parameters",
-                        self._current(),
-                    )
-                self._advance()  # Consume '⟩' or '>'
-
-        return FreeBranch(
-            name=branch_name,
-            parameters=parameters,
-            line=branch_token.line,
-            column=branch_token.column,
-        )
-
-    def _parse_free_type(self) -> FreeType:
-        """Parse free type: Type ::= branch1 | branch2⟨N⟩ | branch3⟨Tree x Tree⟩.
-
-        Supports recursive constructors with parameters.
-
-        Examples:
-            Status ::= active | inactive (simple branches)
-            Tree ::= stalk | leaf⟨N⟩ | branch⟨Tree x Tree⟩ (with parameters)
-        """
-        # Identifier already consumed in _parse_document_item, need to back up
-        name_token = self._current()
-        type_name = name_token.value
-        self._advance()  # Move to ::=
-
-        if not self._match(TokenType.FREE_TYPE):
-            raise ParserError("Expected '::=' in free type definition", self._current())
-        self._advance()  # Consume '::='
-
-        branches: list[FreeBranch] = []
-
-        # Parse pipe-separated list of branches
-        # Stop at NEWLINE, END (for zed blocks), or EOF
-        while not self._at_end() and not self._match(TokenType.NEWLINE, TokenType.END):
-            # Accept IDENTIFIER or keywords that could be branch names (P, F, etc.)
-            if (
-                self._match(TokenType.IDENTIFIER)
-                or self._match(TokenType.POWER)
-                or self._match(TokenType.POWER1)
-                or self._match(TokenType.FINSET)
-                or self._match(TokenType.FINSET1)
-            ):
-                branch_token = self._current()
-                self._advance()
-                branches.append(self._parse_free_branch(branch_token))
-
-            # Check for pipe separator
-            if self._match(TokenType.PIPE):
-                self._advance()
-            elif (
-                not self._match(TokenType.NEWLINE, TokenType.END) and not self._at_end()
-            ):
-                # Unexpected token - raise error to prevent infinite loop
-                current = self._current()
-                if current.type == TokenType.EQUALS:
-                    raise ParserError(
-                        "Unexpected '=' in free type definition. "
-                        "Did you mean '::=' instead of '::=='?",
-                        current,
-                    )
-                raise ParserError(
-                    f"Expected branch name or '|' in free type definition, "
-                    f"got {current.type.name}",
-                    current,
-                )
-
-        if not branches:
-            raise ParserError(
-                "Expected at least one branch in free type definition", self._current()
-            )
-
-        return FreeType(
-            name=type_name,
-            branches=branches,
-            line=name_token.line,
-            column=name_token.column,
-        )
-
-    def _parse_syntax_block(self) -> SyntaxBlock:
-        """Parse syntax environment for aligned free type definitions.
-
-        Syntax:
-          syntax
-            TypeName ::= branch1 | branch2
-
-            AnotherType ::= branch1<Param>
-                         |  branch2<Param1 cross Param2>
-          end
-
-        Generates column-aligned LaTeX with \also between groups.
-        """
-        start_token = self._advance()  # Consume 'syntax'
-        self._skip_newlines()
-
-        # Parse groups of definitions separated by blank lines
-        groups: list[list[SyntaxDefinition]] = []
-        current_group: list[SyntaxDefinition] = []
-
-        while not self._at_end() and not self._match(TokenType.END):
-            # Check for blank line (group separator)
-            if self._match(TokenType.NEWLINE):
-                # Count consecutive newlines
-                newline_count = 0
-                while self._match(TokenType.NEWLINE):
-                    newline_count += 1
-                    self._advance()
-
-                # Blank line (2+ newlines) separates groups
-                if newline_count >= 2 and current_group:
-                    groups.append(current_group)
-                    current_group = []
-
-                # Skip any remaining newlines
-                self._skip_newlines()
-
-                if self._match(TokenType.END):
-                    break
-
-            # Parse a single free type definition
-            if not self._match(TokenType.IDENTIFIER):
-                if not self._match(TokenType.END):
-                    current_type = self._current().type.name
-                    raise ParserError(
-                        f"Expected type name in syntax block, got {current_type}",
-                        self._current(),
-                    )
-                break
-
-            # Parse: TypeName ::= branches
-            type_name_token = self._current()
-            type_name = type_name_token.value
-            self._advance()
-
-            if not self._match(TokenType.FREE_TYPE):
-                raise ParserError(
-                    "Expected '::=' after type name in syntax block",
-                    self._current(),
-                )
-            self._advance()  # Consume '::='
-
-            # Parse branches using shared helper
-            branches: list[FreeBranch] = []
-
-            # Parse initial set of branches on the same line
-            while not self._at_end():
-                if self._match(TokenType.NEWLINE, TokenType.END):
-                    break
-                # Parse branch name
-                if not self._match(TokenType.IDENTIFIER):
-                    if self._match(TokenType.PIPE):
-                        # Skip leading pipe if present
-                        self._advance()
-                        continue
-                    break
-
-                branch_token = self._current()
-                self._advance()
-                branches.append(self._parse_free_branch(branch_token))
-
-                # Check for pipe separator
-                if self._match(TokenType.PIPE):
-                    self._advance()
-
-            # Check for continuation lines (starting with |)
-            while not self._at_end() and not self._match(TokenType.END):
-                # Skip single newline
-                if self._match(TokenType.NEWLINE):
-                    self._advance()
-
-                # Check if next line starts with |
-                if not self._match(TokenType.PIPE):
-                    break
-
-                self._advance()  # Consume continuation '|'
-
-                # Parse branches on this continuation line
-                while not self._at_end():
-                    if self._match(TokenType.NEWLINE, TokenType.END):
-                        break
-
-                    if not self._match(TokenType.IDENTIFIER):
-                        if self._match(TokenType.PIPE):
-                            self._advance()
-                            continue
-                        break
-
-                    branch_token = self._current()
-                    self._advance()
-                    branches.append(self._parse_free_branch(branch_token))
-
-                    if self._match(TokenType.PIPE):
-                        self._advance()
-
-            if not branches:
-                raise ParserError(
-                    f"Expected at least one branch for type {type_name}",
-                    type_name_token,
-                )
-
-            # Create SyntaxDefinition
-            current_group.append(
-                SyntaxDefinition(
-                    name=type_name,
-                    branches=branches,
-                    line=type_name_token.line,
-                    column=type_name_token.column,
-                )
-            )
-
-            # Skip trailing newline after definition
-            if self._match(TokenType.NEWLINE):
-                self._advance()
-
-        # Add final group if not empty
-        if current_group:
-            groups.append(current_group)
-
-        # Expect 'end'
-        if not self._match(TokenType.END):
-            raise ParserError("Expected 'end' to close syntax block", self._current())
-        self._advance()  # Consume 'end'
-
-        if not groups:
-            raise ParserError(
-                "Expected at least one type definition in syntax block", start_token
-            )
-
-        return SyntaxBlock(
-            groups=groups,
-            line=start_token.line,
-            column=start_token.column,
-        )
-
     def _scan_for_defs_after_brackets(self, offset: int) -> bool:
         """Return True when a balanced bracket group starting at ``offset`` is
         followed immediately by a DEFS token.
@@ -5303,376 +4631,6 @@ class Parser:
                 depth -= 1
             offset += 1
         return self._peek_ahead(offset).type == TokenType.DEFS
-
-    def _parse_horiz_def_generic_params(self) -> list[str] | None:
-        """Parse optional generic LHS parameters for a horizontal definition.
-
-        Identical to ``_parse_generic_params`` but records the opening bracket
-        token for an improved error message when the bracket is not closed.
-
-        Returns None when no ``[`` is present.
-        """
-        if not self._match(TokenType.LBRACKET):
-            return None
-
-        open_tok = self._advance()  # Consume '[', save for error reporting
-        params: list[str] = []
-
-        while not self._match(TokenType.RBRACKET) and not self._at_end():
-            if self._match(TokenType.COMMA):
-                self._advance()  # Skip comma
-                continue
-
-            if self._match(TokenType.NEWLINE, TokenType.EOF):
-                raise ParserError(
-                    "unclosed '[' in generic parameter list for horizontal definition",
-                    open_tok,
-                )
-
-            if not self._match(TokenType.IDENTIFIER):
-                raise ParserError(
-                    "expected type parameter name in generic list",
-                    self._current(),
-                )
-
-            params.append(self._current().value)
-            self._advance()
-
-        if not self._match(TokenType.RBRACKET):
-            raise ParserError(
-                "unclosed '[' in generic parameter list for horizontal definition",
-                open_tok,
-            )
-        self._advance()  # Consume ']'
-
-        return params if params else None
-
-    # ------------------------------------------------------------------
-    # Schema-calculus operators (Phase 3.2 — Z RM §3.11)
-    # ------------------------------------------------------------------
-    # Precedence (tightest to loosest):
-    #   1. hide / project   (_parse_schema_project_hide)
-    #   2. composition ;    (_parse_schema_compose)
-    #   3. piping >>        (_parse_schema_pipe)
-    #
-    # These methods are only entered from _parse_horiz_def_rhs (and any
-    # future schema-expression context) when _in_schema_expr_context is
-    # True.  Inside axdef / schema / gendef bodies SEMICOLON remains a
-    # plain declaration separator — those loops never call these methods.
-    # ------------------------------------------------------------------
-
-    def _parse_schema_pipe(self) -> Expr:
-        """Parse schema piping ``S >> T`` (lowest schema-calculus precedence).
-
-        Left-associative: ``A >> B >> C`` parses as ``(A >> B) >> C``.
-        """
-        left = self._parse_schema_compose()
-
-        while self._match(TokenType.PIPE_PIPE):
-            op_tok = self._advance()
-            right = self._parse_schema_compose()
-            left = SchemaPipe(
-                left=left,
-                right=right,
-                line=op_tok.line,
-                column=op_tok.column,
-            )
-
-        return left
-
-    def _parse_schema_compose(self) -> Expr:
-        """Parse schema composition ``S ; T`` (middle schema-calculus precedence).
-
-        Left-associative: ``A ; B ; C`` parses as ``(A ; B) ; C``.
-        SEMICOLON is consumed here only because _in_schema_expr_context is True;
-        callers in declaration-loop contexts never invoke this method.
-        """
-        left = self._parse_schema_project_hide()
-
-        while self._match(TokenType.SEMICOLON):
-            op_tok = self._advance()
-            right = self._parse_schema_project_hide()
-            left = SchemaCompose(
-                left=left,
-                right=right,
-                line=op_tok.line,
-                column=op_tok.column,
-            )
-
-        return left
-
-    def _parse_schema_project_hide(self) -> Expr:
-        """Parse schema hiding and projection (tightest schema-calculus precedence).
-
-        ``S hide (x, y)``  — existentially quantify named components.
-        ``S project T``    — project S onto the signature of T.
-
-        Both are left-associative and at the same precedence level, so they
-        chain: ``S hide (x) project T`` is ``(S hide (x)) project T``.
-        """
-        # Start with a base schema expression (ordinary expression precedence).
-        left = self._parse_expr()
-
-        while self._match(TokenType.HIDE, TokenType.PROJECT):
-            op_tok = self._advance()
-
-            if op_tok.type == TokenType.HIDE:
-                # Expect '(' name, name, ... ')'
-                if not self._match(TokenType.LPAREN):
-                    raise ParserError(
-                        "expected '(' after 'hide' in schema hiding expression",
-                        self._current(),
-                    )
-                self._advance()  # Consume '('
-
-                names: list[str] = []
-                while not self._match(TokenType.RPAREN):
-                    if self._at_end():
-                        raise ParserError(
-                            "unclosed '(' in 'hide' expression — expected ')'",
-                            op_tok,
-                        )
-                    if self._match(TokenType.COMMA):
-                        self._advance()
-                        continue
-                    if not self._match(TokenType.IDENTIFIER):
-                        cur = self._current()
-                        raise ParserError(
-                            f"expected identifier in 'hide' name list,"
-                            f" got {cur.type.name} ({cur.value!r})",
-                            cur,
-                        )
-                    names.append(self._advance().value)
-
-                if not names:
-                    raise ParserError(
-                        "'hide' requires at least one component name", op_tok
-                    )
-                self._advance()  # Consume ')'
-
-                left = SchemaHide(
-                    schema=left,
-                    names=names,
-                    line=op_tok.line,
-                    column=op_tok.column,
-                )
-
-            else:  # PROJECT
-                right = self._parse_expr()
-                left = SchemaProject(
-                    left=left,
-                    right=right,
-                    line=op_tok.line,
-                    column=op_tok.column,
-                )
-
-        return left
-
-    def _parse_horiz_def(self) -> HorizDef:
-        """Parse horizontal schema definition: Name [X, Y]? defs RHS.
-
-        Per Z RM §3.8, the RHS is one of:
-        - A schema reference — an identifier, possibly Phase-0 decorated
-          (e.g. ``Counter'``, ``Delta Counter``).
-        - An inline schema text ``[ decl-list | pred-list ]``.
-        - A schema-calculus expression (Phase 3.2): S ; T, S >> T,
-          S hide (x, y), S project T, and combinations.
-        """
-        start_tok = self._current()
-
-        # Consume the LHS name
-        if not self._match(TokenType.IDENTIFIER):
-            raise ParserError(
-                "expected schema name for horizontal definition", start_tok
-            )
-        name_tok = self._advance()
-        name = name_tok.value
-
-        # Optional generic parameters on the LHS: Name[X, Y]
-        generics = self._parse_horiz_def_generic_params()
-
-        # Consume 'defs'
-        if not self._match(TokenType.DEFS):
-            raise ParserError(
-                "expected 'defs' in horizontal schema definition",
-                self._current(),
-            )
-        self._advance()  # Consume 'defs'
-
-        # --- Parse RHS ---
-        if self._at_end() or self._match(TokenType.NEWLINE):
-            raise ParserError(
-                "expected RHS after 'defs' in horizontal schema definition",
-                self._current(),
-            )
-
-        body: Expr | SchemaInclusion
-        if self._match(TokenType.LBRACKET):
-            body = self._parse_schema_text()
-        else:
-            # Enable schema-expression context so ';' becomes composition.
-            prev = self._in_schema_expr_context
-            self._in_schema_expr_context = True
-            try:
-                body = self._parse_horiz_def_rhs()
-            finally:
-                self._in_schema_expr_context = prev
-
-        return HorizDef(
-            name=name,
-            generics=generics,
-            body=body,
-            line=start_tok.line,
-            column=start_tok.column,
-        )
-
-    def _parse_horiz_def_rhs(self) -> Expr | SchemaInclusion:
-        """Parse a schema expression on the RHS of a horizontal definition.
-
-        Accepts:
-        - A plain identifier (possibly Phase-0 decorated).
-        - ``Delta Name [generics]?`` or ``Xi Name [generics]?`` — schema
-          inclusions used as the RHS (legal in Z RM §3.8 examples).
-        - A generic instantiation: ``Name[T]`` produced by ``_parse_atom``.
-        - A schema-calculus expression (Phase 3.2): ``S ; T``, ``S >> T``,
-          ``S hide (x, y)``, ``S project T``, and combinations.
-
-        Returns an Expr node or SchemaInclusion for Delta/Xi decorated forms.
-        Delta/Xi are handled first because they begin with reserved keywords
-        that would otherwise stop expression parsing.
-        """
-        # Delta / Xi decorated reference → SchemaInclusion
-        # These are handled before schema-calculus parsing because Delta/Xi
-        # are keywords that mark the decorated form; schema-calculus operators
-        # cannot apply to a bare Delta/Xi keyword start.
-        if self._match(TokenType.DELTA, TokenType.XI):
-            start_tok = self._current()
-            decoration = "delta" if self._match(TokenType.DELTA) else "xi"
-            self._advance()  # Consume Delta / Xi
-            if not self._match(TokenType.IDENTIFIER):
-                cur = self._current()
-                kw = "Delta" if decoration == "delta" else "Xi"
-                raise ParserError(
-                    f"expected schema name after {kw},"
-                    f" got {cur.type.name} ({cur.value!r})",
-                    cur,
-                )
-            name_tok = self._advance()
-            generics = self._parse_inclusion_generic_args()
-            return SchemaInclusion(
-                name=name_tok.value,
-                decoration=decoration,
-                generics=generics,
-                line=start_tok.line,
-                column=start_tok.column,
-            )
-
-        # Schema-calculus expression (possibly just a plain identifier).
-        # _parse_schema_pipe is the entry point of the schema-calculus
-        # precedence cascade (pipe < compose < hide/project < base expr).
-        return self._parse_schema_pipe()
-
-    def _parse_schema_text(self) -> SchemaText:
-        r"""Parse an inline schema text: ``[ decl-list | pred-list ]``.
-
-        The bracket has already been detected but NOT consumed when this
-        method is called.  Grammar:
-
-            schema_text ::= '[' decl (';' decl)* '|' pred (';' pred)* ']'
-
-        The declaration forms are the same three supported inside schema and
-        axdef bodies.  Multiple predicates are separated by ``;``.
-        """
-        open_tok = self._advance()  # Consume '['
-
-        # --- Declaration section ---
-        declarations: list[Declaration | SchemaInclusion] = []
-
-        while not self._at_end() and not self._match(
-            TokenType.PIPE, TokenType.RBRACKET
-        ):
-            if self._match(TokenType.NEWLINE):
-                self._advance()
-                continue
-            if self._match(TokenType.SEMICOLON):
-                self._advance()  # Declaration separator
-                continue
-
-            # pk has no relation name in inline schema text — reject with clear message.
-            if self._match(TokenType.PK):
-                raise ParserError(
-                    "Primary-key annotation requires a named schema;"
-                    " inline schema text is anonymous",
-                    self._current(),
-                )
-
-            if (
-                self._match(TokenType.DELTA, TokenType.XI)
-                or self._match(TokenType.IDENTIFIER)
-                or self._is_keyword_usable_as_identifier()
-            ):
-                result = self._parse_declaration_or_inclusion()
-                if isinstance(result, list):
-                    declarations.extend(result)
-                else:
-                    declarations.append(result)
-            else:
-                cur = self._current()
-                raise ParserError(
-                    f"expected declaration or '|' in inline schema text,"
-                    f" got {cur.type.name} ({cur.value!r})",
-                    cur,
-                )
-
-        # FIX 2: empty declaration list is a syntax error
-        if not declarations:
-            raise ParserError(
-                "inline schema text requires at least one declaration",
-                open_tok,
-            )
-
-        # --- Predicate separator '|' ---
-        if self._at_end():
-            raise ParserError(
-                "unclosed inline schema text '[' — expected '|' then ']'", open_tok
-            )
-        if self._match(TokenType.RBRACKET):
-            # No predicate section — parse as [ decls ] (no pipe)
-            self._advance()  # Consume ']'
-            return SchemaText(
-                declarations=declarations,
-                predicates=[],
-                line=open_tok.line,
-                column=open_tok.column,
-            )
-
-        self._advance()  # Consume '|'
-
-        # --- Predicate section (flat list; ';' and newlines are separators) ---
-        predicates: list[Expr] = []
-
-        while not self._at_end() and not self._match(TokenType.RBRACKET):
-            if self._match(TokenType.NEWLINE):
-                self._advance()
-                continue
-            if self._match(TokenType.SEMICOLON):
-                self._advance()
-                continue
-
-            predicates.append(self._parse_expr())
-
-        if not self._match(TokenType.RBRACKET):
-            raise ParserError(
-                "unclosed inline schema text '[' — expected closing ']'", open_tok
-            )
-        self._advance()  # Consume ']'
-
-        return SchemaText(
-            declarations=declarations,
-            predicates=predicates,
-            line=open_tok.line,
-            column=open_tok.column,
-        )
 
     def _parse_generic_params(self) -> list[str] | None:
         """Parse optional generic parameters: [X, Y, Z].
@@ -5706,662 +4664,3 @@ class Parser:
         self._advance()  # Consume ']'
 
         return params if params else None
-
-    def _parse_compound_identifier_name(self) -> str:
-        """Parse identifier name that may have postfix operator suffix.
-
-        Handles compound identifiers like R+, R*, R~ used in abbreviations
-        and schema names where the postfix operator is part of the name itself,
-        not an operation on the identifier.
-
-        Examples:
-            R+ == {a, b : N | b > a}  # R+ is the abbreviation name
-            R* == R+ o9 R+             # R* is the abbreviation name
-            R~ == inv R                # R~ is the abbreviation name
-
-        Returns:
-            Compound identifier string (e.g., "R+", "R*", "R~")
-        """
-        if not self._match(TokenType.IDENTIFIER):
-            raise ParserError("Expected identifier", self._current())
-
-        name_token = self._advance()
-        name = name_token.value
-
-        # Check for postfix closure operators as part of the name
-        # These are valid in definition contexts (not expression contexts)
-        if self._match(TokenType.PLUS, TokenType.STAR, TokenType.TILDE):
-            op_token = self._advance()
-            name = name + op_token.value  # "R" + "+" → "R+"
-
-        return name
-
-    def _parse_abbreviation(self) -> Abbreviation:
-        """Parse abbreviation: [X] name == expression or name == expression.
-
-        Supports optional generic parameters.
-        """
-        start_token = self._current()
-
-        # Check for generic parameters before name
-        generic_params = self._parse_generic_params()
-
-        # Parse name (may include postfix operator suffix like R+, R*, R~)
-        name_token = self._current()  # Save for line/column info
-        name = self._parse_compound_identifier_name()
-
-        if not self._match(TokenType.ABBREV):
-            raise ParserError("Expected '==' in abbreviation", self._current())
-        self._advance()  # Consume '=='
-
-        # Parse expression in relational context: a top-level abbreviation RHS
-        # is an expression, not a Z paragraph body, so a postfix `R[a/b]` is a
-        # relational rename (RelationRename → inline math) rather than a
-        # schema rename (SchemaRename, which fuzz rejects when emitted with a
-        # `/` inside a Z paragraph).
-        prev_relational = self._in_relational_context
-        self._in_relational_context = True
-        try:
-            expr = self._parse_expr()
-        finally:
-            self._in_relational_context = prev_relational
-
-        return Abbreviation(
-            name=name,
-            expression=expr,
-            generic_params=generic_params,
-            line=start_token.line if generic_params else name_token.line,
-            column=start_token.column if generic_params else name_token.column,
-        )
-
-    def _parse_axdef(self) -> AxDef:
-        """Parse axiomatic definition block.
-
-        Syntax: axdef [X, Y] ... end
-        Supports optional generic parameters and semicolon-separated declarations.
-        Comma-separated variable lists share a type: var1, var2 : Type.
-        Schema inclusions (bare, Delta, Xi) are also permitted.
-        """
-        start_token = self._advance()  # Consume 'axdef'
-        self._skip_newlines()
-
-        # Check for generic parameters after 'axdef'
-        generic_params = self._parse_generic_params()
-        self._skip_newlines()
-
-        declarations: list[Declaration | SchemaInclusion] = []
-
-        # Parse declarations until 'where' or 'end'
-        while not self._at_end() and not self._match(TokenType.WHERE, TokenType.END):
-            if self._match(TokenType.NEWLINE):
-                self._advance()
-                continue
-
-            # pk is only valid in schema bodies — reject it here with a clear message.
-            if self._match(TokenType.PK):
-                raise ParserError(
-                    "Primary-key annotation is only valid in schema bodies,"
-                    " not inside axdef",
-                    self._current(),
-                )
-
-            # Three forms: Delta/Xi inclusion, bare inclusion, typed decl.
-            if (
-                self._match(TokenType.DELTA, TokenType.XI)
-                or self._match(TokenType.IDENTIFIER)
-                or self._is_keyword_usable_as_identifier()
-            ):
-                result = self._parse_declaration_or_inclusion()
-                if isinstance(result, list):
-                    declarations.extend(result)
-                else:
-                    declarations.append(result)
-
-                # Semicolons are only valid after typed declarations.
-                if self._match(TokenType.SEMICOLON):
-                    self._advance()  # Consume ';'
-                    self._skip_newlines()
-                else:
-                    self._skip_newlines()
-            else:
-                break
-
-        # Parse 'where' clause (optional)
-        predicate_groups: list[list[Expr]] = []  # Default: no groups
-        if self._match(TokenType.WHERE):
-            self._advance()  # Consume 'where'
-            self._skip_newlines()
-
-            # Parse predicates grouped by blank lines
-            predicate_groups = self._parse_predicate_groups((TokenType.END,))
-
-        # Expect 'end'
-        if not self._match(TokenType.END):
-            raise ParserError("Expected 'end' to close axdef block", self._current())
-        self._advance()  # Consume 'end'
-
-        return AxDef(
-            declarations=declarations,
-            predicates=predicate_groups,
-            generic_params=generic_params,
-            line=start_token.line,
-            column=start_token.column,
-        )
-
-    def _parse_gendef(self) -> GenDef:
-        """Parse generic definition block.
-
-        Generic definitions require generic parameters.
-        Syntax: gendef [X, Y] ... end
-        Supports semicolon-separated declarations: f : X -> Y; g : X -> Y
-        """
-        start_token = self._advance()  # Consume 'gendef'
-        self._skip_newlines()
-
-        # Generic parameters are REQUIRED for gendef
-        generic_params = self._parse_generic_params()
-        if not generic_params:
-            raise ParserError(
-                "Generic parameters required for gendef (e.g., gendef [X, Y])",
-                self._current(),
-            )
-        self._skip_newlines()
-
-        declarations: list[Declaration | SchemaInclusion] = []
-
-        # Parse declarations until 'where' or 'end'
-        while not self._at_end() and not self._match(TokenType.WHERE, TokenType.END):
-            if self._match(TokenType.NEWLINE):
-                self._advance()
-                continue
-
-            # pk is only valid in schema bodies — reject it here with a clear message.
-            if self._match(TokenType.PK):
-                raise ParserError(
-                    "Primary-key annotation is only valid in schema bodies,"
-                    " not inside gendef",
-                    self._current(),
-                )
-
-            # Three forms: Delta/Xi inclusion, bare inclusion, typed declaration.
-            if (
-                self._match(TokenType.DELTA, TokenType.XI)
-                or self._match(TokenType.IDENTIFIER)
-                or self._is_keyword_usable_as_identifier()
-            ):
-                result = self._parse_declaration_or_inclusion()
-                if isinstance(result, list):
-                    declarations.extend(result)
-                else:
-                    declarations.append(result)
-
-                if self._match(TokenType.SEMICOLON):
-                    self._advance()  # Consume ';'
-                    self._skip_newlines()
-                else:
-                    self._skip_newlines()
-            else:
-                break
-
-        # Parse 'where' clause (optional)
-        predicate_groups: list[list[Expr]] = []  # Default: no groups
-        if self._match(TokenType.WHERE):
-            self._advance()  # Consume 'where'
-            self._skip_newlines()
-
-            # Parse predicates grouped by blank lines
-            predicate_groups = self._parse_predicate_groups((TokenType.END,))
-
-        # Expect 'end'
-        if not self._match(TokenType.END):
-            raise ParserError("Expected 'end' to close gendef block", self._current())
-        self._advance()  # Consume 'end'
-
-        return GenDef(
-            generic_params=generic_params,
-            declarations=declarations,
-            predicates=predicate_groups,
-            line=start_token.line,
-            column=start_token.column,
-        )
-
-    def _parse_zed(self) -> Zed:
-        """Parse zed block for Z notation constructs.
-
-        Zed blocks contain unboxed Z notation paragraphs (\\begin{zed}...\\end{zed}).
-        Syntax: zed <content> end
-
-        The content can be:
-        - Given types: given A, B, C
-        - Free types: Type ::= branch1 | branch2
-        - Abbreviations: Name == expression
-        - Predicates: forall x : N | x >= 0
-
-        Multiple constructs can appear in one zed block (mixed content).
-        Single predicates are parsed as expressions (backward compatible).
-        """
-        start_token = self._advance()  # Consume 'zed'
-        self._skip_newlines()
-
-        items: list[DocumentItem] = []
-
-        # Parse multiple statements until 'end'
-        while not self._at_end() and not self._match(TokenType.END):
-            if self._match(TokenType.NEWLINE):
-                self._advance()
-                continue
-
-            # Check for given types
-            if self._match(TokenType.GIVEN):
-                items.append(self._parse_given_type())
-                self._skip_newlines()
-                continue
-
-            # Check for abbreviations with generic parameters [X] Name == ...
-            if self._match(TokenType.LBRACKET):
-                items.append(self._parse_abbreviation())
-                self._skip_newlines()
-                continue
-
-            # Check for free types or abbreviations (both start with IDENTIFIER)
-            if self._match(TokenType.IDENTIFIER):
-                # Save position to potentially backtrack
-                saved_pos = self.pos
-
-                # Try to parse as compound identifier (handles R, R+, R*, R~, etc.)
-                try:
-                    _ = self._parse_compound_identifier_name()
-
-                    # Check what follows the identifier
-                    if self._match(TokenType.FREE_TYPE):
-                        # It's a free type, backtrack and parse properly
-                        self.pos = saved_pos
-                        items.append(self._parse_free_type())
-                    elif self._match(TokenType.ABBREV):
-                        # It's an abbreviation, backtrack and parse properly
-                        self.pos = saved_pos
-                        items.append(self._parse_abbreviation())
-                    else:
-                        # Not a recognized Z notation construct, backtrack and
-                        # parse as expression
-                        self.pos = saved_pos
-                        items.append(self._parse_expr())
-                except ParserError:
-                    # Failed to parse compound identifier, backtrack and
-                    # parse as expression
-                    self.pos = saved_pos
-                    items.append(self._parse_expr())
-
-                self._skip_newlines()
-                continue
-
-            # Otherwise, parse as expression (predicate)
-            items.append(self._parse_expr())
-            self._skip_newlines()
-            break  # Single expression mode (backward compatible)
-
-        # Expect 'end'
-        if not self._match(TokenType.END):
-            raise ParserError("Expected 'end' to close zed block", self._current())
-        self._advance()  # Consume 'end'
-
-        # If single item and it's an expression, return Zed(content=expr)
-        # for backward compatibility. Otherwise, wrap in Document
-        if len(items) == 1 and isinstance(items[0], Expr):
-            return Zed(
-                content=items[0],
-                line=start_token.line,
-                column=start_token.column,
-            )
-        # Multiple items or non-expression items - wrap in Document
-        doc = Document(
-            items=items,
-            line=start_token.line,
-            column=start_token.column,
-        )
-        return Zed(
-            content=doc,
-            line=start_token.line,
-            column=start_token.column,
-        )
-
-    def _parse_schema(self) -> Schema:
-        """Parse schema definition block.
-
-        Syntax:
-            schema Name[X, Y] ... end (named with generics)
-            schema Name ... end (named)
-            schema ... end (anonymous)
-
-        Supports optional generic parameters and semicolon-separated declarations.
-        """
-        start_token = self._advance()  # Consume 'schema'
-
-        # Parse optional schema name
-        # Note: May include postfix operator suffix like S+, S*, S~
-        # (partial support, GitHub #3 open)
-        name: str | None = None
-        generic_params: list[str] | None = None
-
-        if self._match(TokenType.IDENTIFIER):
-            name = self._parse_compound_identifier_name()
-            # Check for generic parameters after name
-            generic_params = self._parse_generic_params()
-
-        self._skip_newlines()
-
-        declarations: list[Declaration | SchemaInclusion] = []
-
-        # Parse declarations until 'where' or 'end'
-        while not self._at_end() and not self._match(TokenType.WHERE, TokenType.END):
-            if self._match(TokenType.NEWLINE):
-                self._advance()
-                continue
-
-            # pk requires a named schema — reject it in anonymous schemas.
-            if self._match(TokenType.PK) and name is None:
-                raise ParserError(
-                    "Primary-key annotation requires a named schema;"
-                    " anonymous schemas have no relation name for PK notation",
-                    self._current(),
-                )
-
-            # Four forms: pk prefix (named schemas only), Delta/Xi inclusion,
-            # bare inclusion, typed declaration.
-            if (
-                self._match(TokenType.PK)
-                or self._match(TokenType.DELTA, TokenType.XI)
-                or self._match(TokenType.IDENTIFIER)
-                or self._is_keyword_usable_as_identifier()
-            ):
-                result = self._parse_declaration_or_inclusion()
-                if isinstance(result, list):
-                    declarations.extend(result)
-                else:
-                    declarations.append(result)
-
-                if self._match(TokenType.SEMICOLON):
-                    self._advance()  # Consume ';'
-                    self._skip_newlines()
-                else:
-                    self._skip_newlines()
-            else:
-                break
-
-        # Parse 'where' clause (optional)
-        predicate_groups: list[list[Expr]] = []  # Default: no groups
-        if self._match(TokenType.WHERE):
-            self._advance()  # Consume 'where'
-            self._skip_newlines()
-
-            # Parse predicates grouped by blank lines
-            predicate_groups = self._parse_predicate_groups((TokenType.END,))
-
-        # Expect 'end'
-        if not self._match(TokenType.END):
-            raise ParserError("Expected 'end' to close schema block", self._current())
-        self._advance()  # Consume 'end'
-
-        return Schema(
-            name=name,
-            declarations=declarations,
-            predicates=predicate_groups,
-            generic_params=generic_params,
-            line=start_token.line,
-            column=start_token.column,
-        )
-
-    def _parse_proof_tree(self) -> ProofTree:
-        """Parse proof tree with Path C syntax (conclusion with supporting proof).
-
-        Supports top-level CASE analysis where proof starts with cases.
-        """
-        start_token = self._advance()  # Consume 'PROOF:'
-        self._skip_newlines()
-
-        if self._at_end() or self._is_structural_token():
-            raise ParserError("Expected proof node after PROOF:", self._current())
-
-        # Check if first item is a case keyword (top-level case analysis)
-        if self._match(TokenType.IDENTIFIER) and self._current().value == "case":
-            # Parse all top-level cases
-            cases: list[CaseAnalysis] = []
-
-            while (
-                not self._at_end()
-                and not self._is_structural_token()
-                and self._match(TokenType.IDENTIFIER)
-                and self._current().value == "case"
-            ):
-                case_analysis = self._parse_case_analysis(
-                    base_indent=0, parent_indent=None
-                )
-                cases.append(case_analysis)
-                self._skip_newlines()
-
-            # Check if there's a final conclusion after the cases
-            final_conclusion_expr: Expr | None = None
-            final_justification: str | None = None
-
-            if not self._at_end() and not self._is_structural_token():
-                # Parse potential final conclusion (e.g., "q [or elim]")
-                final_conclusion_node = self._parse_proof_node(
-                    base_indent=0, parent_indent=None
-                )
-                final_conclusion_expr = final_conclusion_node.expression
-                final_justification = final_conclusion_node.justification
-
-            # Create synthetic conclusion node to wrap the cases
-            # Use the final conclusion if present, otherwise placeholder
-            if final_conclusion_expr is None:
-                final_conclusion_expr = Identifier(
-                    name="[case_analysis]",
-                    line=start_token.line,
-                    column=start_token.column,
-                )
-                final_justification = "case analysis"
-
-            # Type hint for children: explicitly cast to expected union type
-            children_list: list[ProofNode | CaseAnalysis] = list(cases)
-
-            conclusion = ProofNode(
-                expression=final_conclusion_expr,
-                justification=final_justification,
-                label=None,
-                is_assumption=False,
-                is_sibling=False,
-                children=children_list,
-                indent_level=0,
-                line=start_token.line,
-                column=start_token.column,
-            )
-        else:
-            # Standard proof: parse the conclusion node
-            conclusion = self._parse_proof_node(base_indent=0, parent_indent=None)
-
-        return ProofTree(
-            conclusion=conclusion, line=start_token.line, column=start_token.column
-        )
-
-    def _parse_proof_node(
-        self, base_indent: int, parent_indent: int | None
-    ) -> ProofNode:
-        """Parse a single proof node with Path C syntax."""
-        if self._at_end() or self._is_structural_token():
-            raise ParserError("Expected proof node", self._current())
-
-        line_token = self._current()
-        current_indent = line_token.column
-
-        # If we've dedented past the parent level, stop
-        if parent_indent is not None and current_indent <= parent_indent:
-            raise ParserError("Unexpected dedent in proof tree", self._current())
-
-        # Check for assumption label [1], [2], etc.
-        label: int | None = None
-        if self._match(TokenType.LBRACKET):
-            self._advance()  # Consume '['
-            if self._match(TokenType.NUMBER):
-                label = int(self._current().value)
-                self._advance()
-                if not self._match(TokenType.RBRACKET):
-                    raise ParserError(
-                        "Expected ']' after assumption label", self._current()
-                    )
-                self._advance()  # Consume ']'
-            else:
-                # Not a label, must be justification - back up by recreating the bracket
-                # Actually, we can't back up easily. Let's handle this differently.
-                # For now, assume [number] is always a label at the start of a line.
-                raise ParserError(
-                    "Expected number in assumption label", self._current()
-                )
-
-        # Check for sibling marker ::
-        is_sibling = False
-        if self._match(TokenType.DOUBLE_COLON):
-            is_sibling = True
-            self._advance()  # Consume '::'
-
-        # Check for ellipsis ... (steps omitted in proof)
-        if self._match(TokenType.ELLIPSIS):
-            ellipsis_token = self._advance()
-            # Create a text node representing omitted steps
-            expr: Expr = Identifier(
-                name="...",
-                line=ellipsis_token.line,
-                column=ellipsis_token.column,
-            )
-        else:
-            # Parse the expression
-            expr = self._parse_expr()
-
-        # Check for justification [rule name]
-        justification: str | None = None
-        is_assumption = False
-        if self._match(TokenType.LBRACKET):
-            self._advance()  # Consume '['
-
-            # Collect justification text
-            just_parts: list[str] = []
-            while not self._match(TokenType.RBRACKET) and not self._at_end():
-                if self._match(TokenType.NEWLINE):
-                    break
-                just_parts.append(self._current().value)
-                self._advance()
-
-            if self._match(TokenType.RBRACKET):
-                self._advance()  # Consume ']'
-                # Join tokens smartly: add spaces only between consecutive identifiers
-                justification = self._smart_join_justification(just_parts)
-                # Check if this is the [assumption] keyword
-                if justification == "assumption":
-                    is_assumption = True
-
-        self._skip_newlines()
-
-        # Parse children (nodes indented more than current)
-        children: list[ProofNode | CaseAnalysis] = []
-        while not self._at_end() and not self._is_structural_token():
-            if self._match(TokenType.NEWLINE):
-                self._skip_newlines()
-                if self._at_end() or self._is_structural_token():
-                    break
-
-            next_token = self._current()
-            next_indent = next_token.column
-
-            # Check if still indented relative to current node
-            if next_indent <= current_indent:
-                break
-
-            # Check for case analysis
-            if self._match(TokenType.IDENTIFIER) and self._current().value == "case":
-                case_analysis = self._parse_case_analysis(
-                    base_indent=base_indent, parent_indent=current_indent
-                )
-                children.append(case_analysis)
-            else:
-                # Regular proof node
-                child = self._parse_proof_node(
-                    base_indent=base_indent, parent_indent=current_indent
-                )
-                children.append(child)
-
-        return ProofNode(
-            expression=expr,
-            justification=justification,
-            label=label,
-            is_assumption=is_assumption,
-            is_sibling=is_sibling,
-            children=children,
-            indent_level=current_indent - base_indent,
-            line=line_token.line,
-            column=line_token.column,
-        )
-
-    def _parse_case_analysis(
-        self, base_indent: int, parent_indent: int | None
-    ) -> CaseAnalysis:
-        """Parse case analysis: case name: followed by proof steps.
-
-        parent_indent can be None for top-level cases.
-        """
-        if not self._match(TokenType.IDENTIFIER) or self._current().value != "case":
-            raise ParserError("Expected 'case' keyword", self._current())
-
-        case_token = self._advance()  # Consume 'case'
-
-        # Parse case name - collect all tokens until colon
-        case_parts: list[str] = []
-        while not self._match(TokenType.COLON) and not self._at_end():
-            if self._match(TokenType.NEWLINE):
-                raise ParserError("Expected ':' after case name", self._current())
-            case_parts.append(self._current().value)
-            self._advance()
-
-        if not case_parts:
-            raise ParserError("Expected case name after 'case'", self._current())
-
-        case_name = " ".join(case_parts)
-
-        # Expect colon
-        if not self._match(TokenType.COLON):
-            raise ParserError("Expected ':' after case name", self._current())
-        self._advance()  # Consume ':'
-
-        self._skip_newlines()
-
-        # Parse proof steps indented under this case
-        steps: list[ProofNode] = []
-        case_indent = case_token.column
-
-        while not self._at_end() and not self._is_structural_token():
-            if self._match(TokenType.NEWLINE):
-                self._skip_newlines()
-                if self._at_end() or self._is_structural_token():
-                    break
-
-            next_token = self._current()
-            next_indent = next_token.column
-
-            # Check if still indented relative to case
-            if next_indent <= case_indent:
-                break
-
-            # Check if this is another case (sibling case)
-            if self._match(TokenType.IDENTIFIER) and self._current().value == "case":
-                break  # Stop parsing this case
-
-            # Parse proof node for this case
-            step = self._parse_proof_node(
-                base_indent=base_indent, parent_indent=case_indent
-            )
-            steps.append(step)
-
-        return CaseAnalysis(
-            case_name=case_name,
-            steps=steps,
-            line=case_token.line,
-            column=case_token.column,
-        )
