+++ src/txt2tex/parser_pkg/__init__.py
@@ -1,3 +1,9 @@
 """Parser subpackage — construct-family rule modules for Parser."""
 
 from __future__ import annotations
+
+from txt2tex.parser_pkg import (
+    paragraphs,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+    proofs,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+    schemas,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+)
