+++ src/txt2tex/parser_pkg/_base.py
@@ -1,0 +1,250 @@
+"""Base class for Parser mixins — declares shared state and helper shapes.
+
+``Parser`` is recursive descent: rule methods call their neighbours through
+``self`` rather than through a dispatch table.  When rule methods move to
+mixin files, those ``self.X`` calls keep working at runtime because the final
+``Parser`` class inherits all the mixins through its MRO.
+
+Type checkers, however, see each mixin file in isolation.  This module
+provides a single base class, :class:`ParserBase`, whose ``TYPE_CHECKING``
+block declares every cross-cutting attribute and helper method that mixin
+methods reference.  The declarations have no runtime effect — the real
+implementations remain on :class:`Parser` (or on whichever mixin holds them
+after later moves) and are reached through the normal MRO.
+
+As subsequent moves consolidate helpers into their own mixin files, the
+declarations migrate out of this module in lockstep.
+"""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+from txt2tex.tokens import (
+    Token,
+    TokenType,
+)
+
+
+class ParserError(Exception):
+    """Raised when parser encounters invalid syntax."""
+
+    # Instance variable type annotations
+    message: str
+    token: Token
+
+    def __init__(self, message: str, token: Token) -> None:
+        """Initialize parser error with token position."""
+        super().__init__(f"Line {token.line}, column {token.column}: {message}")
+        self.message = message
+        self.token = token
+
+
+if TYPE_CHECKING:
+    from txt2tex.ast_nodes import (
+        ArgueChain,
+        AxDef,
+        BibliographyMetadata,
+        BMachine,
+        CaseAnalysis,
+        Contents,
+        Declaration,
+        DocumentItem,
+        Expr,
+        FreeBranch,
+        FreeType,
+        GenDef,
+        GenericInstantiation,
+        GivenType,
+        InfruleBlock,
+        LatexBlock,
+        LineBreak,
+        PageBreak,
+        Paragraph,
+        Part,
+        ProofNode,
+        ProofTree,
+        PureParagraph,
+        RawLatexBlock,
+        SchemaBinding,
+        SchemaInclusion,
+        SchemaRename,
+        SchemaText,
+        Section,
+        Solution,
+        SyntaxBlock,
+        TitleMetadata,
+        TruthTable,
+        Zed,
+    )
+
+
+class ParserBase:
+    """Type-only shape declaration for the composed ``Parser`` class.
+
+    Every mixin in this package inherits from this class so that its methods
+    type-check against the full parser interface.  The class has no runtime
+    attributes or methods of its own — see the ``TYPE_CHECKING`` block below
+    for what it declares to the type checker.
+    """
+
+    if TYPE_CHECKING:
+        # --- Token stream + position state ---
+        tokens: list[Token]
+        pos: int
+        last_token_end_column: int
+        last_token_line: int
+
+        # --- Context flags ---
+        _parsing_schema_text: bool
+        _in_comprehension_body: bool
+        _in_schema_expr_context: bool
+        _in_comparison_rhs: bool
+        _current_quantifier_vars: set[str]
+        _in_relational_context: bool
+
+        # --- Token-cursor helpers (Move 21 → lexer_state.py) ---
+        def _at_end(self) -> bool: ...
+        def _current(self) -> Token: ...
+        def _advance(self) -> Token: ...
+        def _match(self, *types: TokenType) -> bool: ...
+        def _peek_ahead(self, offset: int = 1) -> Token | None: ...
+        def _skip_newlines(self) -> None: ...
+        def _has_blank_line(self) -> bool: ...
+        def _is_keyword_usable_as_identifier(self) -> bool: ...
+        def _is_operand_start(self) -> bool: ...
+        def _is_attr_name_token(self) -> bool: ...
+        def _bracket_contains_slash(self) -> bool: ...
+        def _scan_for_colon_before_newline(self) -> bool: ...
+        def _scan_for_defs_after_brackets(self, offset: int) -> bool: ...
+        def _next_non_newline_is_cross_op(self) -> bool: ...
+        def _reject_stray_slash(self) -> None: ...
+        def _reject_reserved_decl_name(self, tok: Token) -> None: ...
+        def _is_structural_token(self) -> bool: ...
+
+        # --- Expression parser (Move 16) ---
+        def _parse_expr(self) -> Expr: ...
+        def _parse_conditional(self) -> Expr: ...
+        def _parse_iff(self) -> Expr: ...
+        def _parse_implies(self) -> Expr: ...
+        def _parse_implies_rhs(self) -> Expr: ...
+        def _parse_or(self) -> Expr: ...
+        def _parse_and(self) -> Expr: ...
+        def _parse_unary(self) -> Expr: ...
+        def _parse_additive(self) -> Expr: ...
+        def _parse_range(self) -> Expr: ...
+        def _parse_multiplicative(self) -> Expr: ...
+        def _parse_comparison(self) -> Expr: ...
+        def _parse_relation(self) -> Expr: ...
+        def _parse_set_op(self) -> Expr: ...
+        def _parse_union(self) -> Expr: ...
+        def _parse_intersect(self) -> Expr: ...
+        def _parse_cross(self) -> Expr: ...
+        def _parse_postfix(self) -> Expr: ...
+        def _parse_atom(self) -> Expr: ...
+        def _parse_argument_list(self) -> list[Expr]: ...
+        def _parse_parenthesized_expr_or_tuple(self) -> Expr: ...
+        def _parse_set(self) -> Expr: ...
+        def _parse_set_comprehension_from_brace(self) -> Expr: ...
+        def _parse_set_predicate(self) -> Expr: ...
+        def _parse_set_expression(self) -> Expr: ...
+        def _parse_sequence_literal(self) -> Expr: ...
+        def _parse_bag_literal(self) -> Expr: ...
+        def _parse_quantifier(self) -> Expr: ...
+        def _parse_quantifier_continuation(
+            self, quantifier: str, line: int, column: int
+        ) -> Expr: ...
+        def _parse_schema_quantifier_body(
+            self, quant_token: Token, schema_binding: SchemaBinding
+        ) -> Expr: ...
+        def _parse_lambda(self) -> Expr: ...
+        def _parse_binding(self) -> Expr: ...
+        def _parse_binding_component(self) -> tuple[str, Expr]: ...
+        def _should_parse_space_separated_arg(self) -> bool: ...
+        def _try_parse_guarded_cases(self, first_expr: Expr) -> Expr: ...
+
+        # --- Type / algebra parsers (Moves 15, 17) ---
+        def _parse_function_type(self) -> Expr: ...
+        def _parse_restrict(self) -> Expr: ...
+        def _parse_project(self) -> Expr: ...
+        def _parse_relation_rename(self, base: Expr) -> Expr: ...
+        def _parse_generic_instantiation(self, base: Expr) -> GenericInstantiation: ...
+        def _parse_group_rhs(self, relation: Expr, group_tok: Token) -> Expr: ...
+        def _parse_group_regroup_rhs(
+            self, relation: Expr, group_tok: Token
+        ) -> Expr: ...
+        def _parse_group_aggregate_rhs(
+            self, relation: Expr, group_tok: Token
+        ) -> Expr: ...
+        def _parse_aggregator_clause(self) -> object: ...
+        def _parse_ungroup_rhs(self, relation: Expr, ungroup_tok: Token) -> Expr: ...
+
+        # --- Schema parsers (Move 13) ---
+        def _parse_schema_pipe(self) -> Expr: ...
+        def _parse_schema_compose(self) -> Expr: ...
+        def _parse_schema_project_hide(self) -> Expr: ...
+        def _parse_schema_text(self) -> SchemaText: ...
+        def _parse_schema_rename(self, schema: Expr) -> SchemaRename: ...
+        def _parse_schema_rename_or_generic(
+            self, base: Expr
+        ) -> GenericInstantiation | SchemaRename: ...
+        def _parse_horiz_def_generic_params(self) -> list[str] | None: ...
+        def _parse_horiz_def_rhs(self) -> Expr | SchemaInclusion: ...
+        def _parse_schema_body(
+            self,
+        ) -> tuple[list[Declaration | SchemaInclusion], list[list[Expr]]]: ...
+
+        # --- Paragraph parsers (Move 12) ---
+        def _parse_given_type(self) -> GivenType: ...
+        def _parse_free_type(self) -> FreeType: ...
+        def _parse_free_branch(self, branch_token: Token) -> FreeBranch: ...
+        def _parse_syntax_block(self) -> SyntaxBlock: ...
+        def _parse_abbreviation(self) -> object: ...
+        def _parse_axdef(self) -> AxDef: ...
+        def _parse_gendef(self) -> GenDef: ...
+        def _parse_zed(self) -> Zed: ...
+
+        # --- Proof parsers (Move 14) ---
+        def _parse_proof_tree(self) -> ProofTree: ...
+        def _parse_proof_node(
+            self, base_indent: int, parent_indent: int | None
+        ) -> ProofNode: ...
+        def _parse_case_analysis(
+            self, base_indent: int, parent_indent: int | None
+        ) -> CaseAnalysis: ...
+        def _parse_argue_chain(self) -> ArgueChain: ...
+        def _parse_infrule_block(self) -> InfruleBlock: ...
+
+        # --- Misc helpers ---
+        def _parse_generic_params(self) -> list[str] | None: ...
+        def _parse_compound_identifier_name(self) -> str: ...
+        def _parse_declaration_or_inclusion(
+            self,
+        ) -> Declaration | SchemaInclusion | list[Declaration]: ...
+        def _parse_predicate_groups(
+            self, end_tokens: tuple[TokenType, ...]
+        ) -> list[list[Expr]]: ...
+        def _parse_inclusion_generic_args(self) -> list[Expr] | None: ...
+        def _smart_join_justification(self, parts: list[str]) -> str: ...
+        def _parse_title_metadata(self) -> TitleMetadata | None: ...
+
+        # --- Document item parsers (Moves 19, structural) ---
+        def _parse_document_item(self) -> DocumentItem: ...
+        def _parse_section(self) -> Section: ...
+        def _parse_solution(self) -> Solution: ...
+        def _parse_part(self) -> Part: ...
+        def _parse_paragraph(self) -> Paragraph | Part: ...
+        def _parse_paragraph_coalesced(self) -> Paragraph | Part: ...
+        def _parse_pure_paragraph(self) -> PureParagraph: ...
+        def _parse_latex_block(self) -> LatexBlock: ...
+        def _parse_raw_latex_block(self) -> RawLatexBlock: ...
+        def _parse_b_block(self) -> BMachine: ...
+        def _parse_pagebreak(self) -> PageBreak: ...
+        def _parse_linebreak(self) -> LineBreak: ...
+        def _parse_contents(self) -> Contents: ...
+        def _parse_parts_directive(self) -> object: ...
+        def _parse_parts_format(self) -> str: ...
+        def _parse_truth_table(self) -> TruthTable: ...
+        def _parse_identifier_document_item(self) -> DocumentItem: ...
+        def _parse_bracket_document_item(self) -> DocumentItem: ...
+        def _parse_bibliography_metadata(self) -> BibliographyMetadata: ...
